server/pyproject.toml
[tool.poetry]
name = "chrome-ext-api"
version = "0.1.0"
description = "An api for storing and fetching video files recorded by a chrome extension"
authors = ["Nasir Abdulrasheed <mypythtesting@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.103.2"
python-decouple = "^3.8"
python-multipart = "^0.0.6"
openai = "^0.28.1"
pydub = "^0.25.1"
ffmpeg-python = "^0.2.0"
uvicorn = "^0.23.2"
hvac = "^2.2.0"
aio-pika = "^9.4.1"
motor = "^3.4.0"
deepgram-sdk = "^3.4.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"
fabric2 = "^3.2.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
hypothesis = "^6.103.2"

[tool.ruff]
line-length = 89
indent-width = 4
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
